add：各jsファイルを追加し、そのファイルにjsを記載
