+++ app/assets/stylesheets/custom.scss
@@ -685,6 +685,10 @@ a {
   width: 600px;
 }
 
+.edit_page .form_control::placeholder {
+  color: #BBB;
+}
+
 .edit_page .label_name {
   font-weight: bold;
 }
@@ -702,11 +706,11 @@ a {
   margin: 13px 0;
   text-align: center;
   width: 50%;
-  background-color: #777;
+  background-color: #AA0000;
   color: #fff;
 
   &:hover {
-    background-color: #999;
+    background-color: #CC0000;
     transition: 0.3s;
   }
 }
@@ -750,9 +754,11 @@ a {
   -o-transition: .3s ease;
   transition: .3s ease;
 }
+
 .file:hover {
   background-color: #00CCFF;
 }
+
 .file_input {
   position: absolute;
   top: 0;
@@ -761,6 +767,14 @@ a {
   height: 100%;
   opacity: 0;
   cursor: pointer;
+}
+
+.device_inner .file{
+  background-color: #777;
+
+  &:hover {
+    background-color: #999;
+  }
 }
 
 .left_column {
