+++ app/javascript/packs/header_follow.js
@@ -1,0 +1,5 @@
+// ヘッダー追従イベント
+window.addEventListener("scroll", function () {
+  var header = document.querySelector("header");
+  header.classList.toggle("hidden", window.scrollY > 0);
+});
