+++ app/javascript/packs/top_button.js
@@ -1,0 +1,10 @@
+// トップへ戻るボタンのイベント
+document.addEventListener("turbolinks:load", function () {
+  const PageTopBtn = document.getElementById('js_button');
+  PageTopBtn.addEventListener('click', () =>{
+    window.scrollTo({
+      top: 0,
+      behavior: 'smooth'
+    });
+  });
+});
