+++ app/views/layouts/application.html.erb
@@ -20,6 +20,7 @@
       <%= yield %>
       <button class="scroll_top" id="js_button"><i class="fa fa-chevron-up" aria-hidden="true"></i></button>
       <%= render 'layouts/footer' %>
+      <script src="https://unpkg.com/swiper/swiper-bundle.min.js"></script>
     </div>
   </body>
 </html>
