+++ app/views/static_pages/home.html.erb
@@ -60,4 +60,3 @@
     <div class="swiper-pagination"></div>
   </div>
 </div>
-<script src="https://unpkg.com/swiper/swiper-bundle.min.js"></script>
